commit fully working version

+++ lib/misc.c
@@ -20,7 +20,7 @@ void print_cube_vertices(struct coord_3d cube_vertices[VERTEX_COUNT])
     printf("\n \n \n=== c_ube vertices ===\n \n");
     for (int i = 0; i < VERTEX_COUNT; i++)
     {
-        printf("(%lf, %lf, %lf)\n", cube_vertices[i].x, cube_vertices[i].y, cube_vertices[i].z);
+        printf("(%d, %d, %d)\n", (int)cube_vertices[i].x, (int)cube_vertices[i].y, (int)cube_vertices[i].z);
     }
 }
 
@@ -32,7 +32,7 @@ void print_triangles(struct coord_3d triangles[TRIANGLE_COUNT][TRIANGLE_VERTICES
         printf("\nTriangle %d: ", i);
         for (int j = 0; j < TRIANGLE_VERTICES; j++)
         {
-            printf("(%lf, %lf, %lf)", triangles[i][j].x, triangles[i][j].y, triangles[i][j].z);
+            printf("(%d, %d, %d)", (int)triangles[i][j].x, (int)triangles[i][j].y, (int)triangles[i][j].z);
             if (j != 2)
             {
                 printf(", ");
